# backend/src/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cinemate API",
    "description": "Full-stack movie streaming and social recommendation platform. Catalog data from PostgreSQL; rich movie detail, cast, and trailers fetched live from TMDB API.",
    "version": "2.0.0"
  },
  "servers": [
    { "url": "/api", "description": "Reverse-proxied through Nginx" }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "HttpOnly JWT cookie set on login/register. Access token TTL 15 min; refresh token 7 days."
      }
    },
...
  },
  "paths": {
...
    "/users/me/stats": {
      "get": {
        "tags": ["Users"],
        "summary": "Your profile stats",
        "security": [{ "cookieAuth": [] }],
        "responses": { "200": { "description": "{ watched, rated, wishlist, avgRating }" } }
      }
    },
    "/users/me/history": {
      "get": {
        "tags": ["Users"],
        "summary": "Your watch history with progress_s per movie",
        "security": [{ "cookieAuth": [] }],
        "responses": { "200": { "description": "Array of movies with progress_s" } }
      }
    },
    "/users/me/favourites": {
      "get": {
        "tags": ["Users"],
        "summary": "Movies you rated ≥ 8",
        "security": [{ "cookieAuth": [] }],
        "responses": { "200": { "description": "Array of movies" } }
      }
    },
    "/users/me/wishlist": {
      "get": {
        "tags": ["Users"],
        "summary": "Your watchlist",
        "security": [{ "cookieAuth": [] }],
        "responses": { "200": { "description": "Array of movies" } }
      }
    },
    "/users/me/comments": {
      "get": {
        "tags": ["Users"],
        "summary": "Your comments with movie titles",
        "security": [{ "cookieAuth": [] }],
        "responses": { "200": { "description": "Array of comments" } }
      }
    },
    "/users/me/ratings": {
      "get": {
        "tags": ["Users"],
        "summary": "All your movie ratings",
        "security": [{ "cookieAuth": [] }],
        "responses": { "200": { "description": "Array of { tmdb_id, title, score, created_at }" } }
      }
    },
    "/users/me/preferences": {
      "get": {
        "tags": ["Users"],
        "summary": "Get saved preferences (moods, duration, context from Discover wizard)",
        "security": [{ "cookieAuth": [] }],
        "responses": { "200": { "description": "{ preferences: object | null }" } }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Save preferences as JSONB",
        "security": [{ "cookieAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": { "200": { "description": "Saved" } }
      }
    },
    "/users/search": {
      "get": {
        "tags": ["Users"],
        "summary": "Search users by username (case-insensitive partial match)",
        "parameters": [{ "name": "q", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Array of { id, username, avatar_url, follower_count }" } }
      }
    },
...
    "/party": {
      "post": {
        "tags": ["Watch Party"],
        "summary": "Create a watch party room — generates 6-char room code",
        "security": [{ "cookieAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["tmdbId"], "properties": { "tmdbId": { "type": "integer" } } } } } },
        "responses": { "201": { "description": "{ room_code, room_id, movie: { tmdb_id, title } }" } }
      }
    },
    "/party/{code}": {
      "get": {
        "tags": ["Watch Party"],
        "summary": "Get room state, last 100 messages, online members",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "name": "code", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "{ room, messages, online_members }" } }
      },
      "delete": {
        "tags": ["Watch Party"],
        "summary": "End room — broadcasts room_ended to all WS clients (host only)",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "name": "code", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Room ended" } }
      }
    },
...
  }
}